# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/accounts": {
      "get": {
        "operationId": "AdminAccountsController_list_v1",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter records (email or name).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Pagination limit.",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset.",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accounts fetched."
          },
          "401": {
            "description": "Admin authentication required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List customer accounts",
        "tags": [
          "admin-accounts"
        ]
      }
    },
    "/v1/admin/accounts/{accountId}": {
      "get": {
        "operationId": "AdminAccountsController_getOne_v1",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account fetched."
          },
          "401": {
            "description": "Admin authentication required."
          },
          "404": {
            "description": "Account not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get account details by ID",
        "tags": [
          "admin-accounts"
        ]
      }
    },
    "/v1/admin/accounts/{accountId}/notifications": {
      "patch": {
        "operationId": "AdminAccountsController_updateNotifications_v1",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification preferences updated."
          },
          "401": {
            "description": "Admin authentication required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update notification preferences for an account",
        "tags": [
          "admin-accounts"
        ]
      }
    },
    "/v1/admin/accounts/{accountId}/achievements": {
      "get": {
        "operationId": "AdminAccountsController_achievements_v1",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Achievements summary fetched."
          },
          "401": {
            "description": "Admin authentication required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch achievement summary for an account",
        "tags": [
          "admin-accounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Koajo API",
    "description": "API documentation for Koajo services",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste your Koajo access token here"
      }
    },
    "schemas": {
...
      "UpdateNotificationPreferencesDto": {
        "type": "object",
        "properties": {
          "emailNotificationsEnabled": {
            "type": "boolean",
            "description": "Enable or disable general system emails.",
            "example": true
          },
          "transactionNotificationsEnabled": {
            "type": "boolean",
            "description": "Enable or disable transaction-related emails.",
            "example": true
          }
        }
      },
...
    }
  }
}